Move InMemoryIndex to package level
This allows reuse across several test modules.
